Interfaz de carro de compras

+++ app/Repositories/Cart/CartRepositoryInterface.php
@@ -16,7 +16,7 @@ interface CartRepositoryInterface
 
   public function remove(int $wineId): void;
 
-  public function getTotalQuantityForWine(Wine $wine): float;
+  public function getTotalQuantityForWine(Wine $wine): int;
 
   public function getTotalCostForWine(Wine $wine, bool $formatted): float|string;
